Caption/requirements.txt
absl-py==0.11.0
astunparse==1.6.3
blis==0.2.4
branca==0.4.1
cached-property==1.5.2
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cymem==2.0.5
Flask==1.1.2
flatbuffers==1.12
folium==0.11.0
gast==0.3.3
gevent==20.9.0
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
greenlet==0.4.17
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==3.3.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
Markdown==3.3.3
MarkupSafe==1.1.1
mathparse==0.1.2
murmurhash==1.0.5
nltk==3.5
numpy==1.19.3
oauthlib==3.1.0
object-detection==0.1
opt-einsum==3.3.0
packaging==20.8
pandas==1.1.5
Pillow==8.0.1
Pint==0.16.1
plac==0.9.6
preshed==2.0.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyYAML==3.13
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
sklearn==0.0
slim==0.1
spacy==2.1.9
SQLAlchemy==1.3.22
srsly==1.0.5
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
thinc==7.0.8
threadpoolctl==2.1.0
tqdm==4.54.1
typing-extensions==3.7.4.3
urllib3==1.26.2
wasabi==0.8.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.2.0
